+++ source/src/Positive.cpp
@@ -485,7 +485,7 @@ void Ostap::Math::KarlinStudden::updateRoots()
   for ( unsigned short k  = 0 ; k < np ; ++k ) 
   { m_troots [ k + 1 ] = m_sphere2.x2 ( k ) ; }
   //
-  std::reverse     ( m_troots.begin () + 1 , m_troots.end () ) ;
+  // std::reverse     ( m_troots.begin () + 1 , m_troots.end () ) ;
   std::partial_sum ( m_troots.begin ()     , m_troots.end () , m_troots.begin() ) ;
   //
   const long double r_small { Ostap::Math::next_float ( 1.0f , -2 )  } ;